+++ lib/src/constants.dart
@@ -1,1 +1,4 @@
+const MIN_VALUE = 0.0;
+const MAX_VALUE = 1.0;
 const arrowAspectRatio = 0.7;
+const durationFactorInMS = (25000.0 / 1024.0);