+++ lib/src/single_straight_line_painter.dart
@@ -31,7 +31,7 @@ class SingleStraightLinePainter implements SingleLinePainter {
 
   @override
   Duration animationDuration() {
-    int duration = _distanceBetweenPoints * 15000 ~/ 1024;
+    int duration = (_distanceBetweenPoints * durationFactorInMS).round();
     return Duration(milliseconds: duration);
   }