+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.2
+
+- Added **update** method to update values of **LineInfo** Object
+- Added **resetAnimation** method to reset animation
+
 # 1.0.1
 
 - Updated Readme